+++ boundflow/backends/tvm/__init__.py
@@ -1,9 +1,13 @@
 from .interval_linear import IntervalLinearKey, build_interval_linear_module
 from .interval_conv2d import IntervalConv2dKey, build_interval_conv2d_module
+from .relax_interval_linear import build_relax_interval_linear_vm_exec
+from .relax_interval_conv2d import build_relax_interval_conv2d_vm_exec
 
 __all__ = [
     "IntervalLinearKey",
     "build_interval_linear_module",
     "IntervalConv2dKey",
     "build_interval_conv2d_module",
+    "build_relax_interval_linear_vm_exec",
+    "build_relax_interval_conv2d_vm_exec",
 ]